third_party/WebKit/LayoutTests/editing/selection/extend/extend_selection_character_14.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../assert_selection.js"></script><script>
function testIt(dir, offset, direction, expectedText) {
  test(() => assert_selection(
  `<span dir="${dir}">${makeSample(offset)}</span>`,
  selection => selection.modify('extend', direction, 'character'),
  `<span dir="${dir}">${expectedText}</span>`),
  `${dir} ${offset} extend ${direction} character`
  );
}

function makeSample(offset){
  const kSample = '\n Just\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n';
  return kSample.substr(0, offset) + '|' + kSample.substring(offset);
}

// ltr right character
testIt('ltr', 0, 'right', '\n ^J|ust\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 1, 'right', '\n ^J|ust\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 2, 'right', '\n ^J|ust\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 3, 'right', '\n J^u|st\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 4, 'right', '\n Ju^s|t\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 5, 'right', '\n Jus^t|\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 6, 'right', '\n Just^\n| <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 7, 'right', '\n Just\n <span>^t|esting \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 8, 'right', '\n Just\n <span>^t|esting \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 9, 'right', '\n Just\n <^s|pan>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 10, 'right', '\n Just\n <s|pan>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</s|pan>');
testIt('ltr', 11, 'right', '\n Just\n <sp|an>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</sp|an>');
testIt('ltr', 12, 'right', '\n Just\n <spa|n>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</spa|n>');
testIt('ltr', 13, 'right', '\n Just\n <span|>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span|>');
testIt('ltr', 14, 'right', '\n Just\n <span>^t|esting \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 15, 'right', '\n Just\n <span>t^e|sting \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 16, 'right', '\n Just\n <span>te^s|ting \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 17, 'right', '\n Just\n <span>tes^t|ing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 18, 'right', '\n Just\n <span>test^i|ng \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 19, 'right', '\n Just\n <span>testi^n|g \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 20, 'right', '\n Just\n <span>testin^g| \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 21, 'right', '\n Just\n <span>testing^ |\u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 22, 'right', '\n Just\n <span>testing ^\u05E8|\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 23, 'right', '\n Just\n <span>testing \u05E8^\u05E7|</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 24, 'right', '\n Just\n <span>testing \u05E8\u05E7</span>^\n| \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 25, 'right', '\n Just\n <span>testing \u05E8\u05E7<^/|span>\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('ltr', 26, 'right', '\n Just\n <span>testing \u05E8\u05E7span^\n| \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('ltr', 27, 'right', '\n Just\n <span>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('ltr', 28, 'right', '\n Just\n <span>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('ltr', 29, 'right', '\n Just\n <span>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('ltr', 30, 'right', '\n Just\n <span>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('ltr', 31, 'right', '\n Just\n <span>testing \u05E8\u05E7</span>^\n| \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 32, 'right', '\n Just\n <span>testing \u05E8\u05E7</span>\n ^\u05D1|\u05D5\u05D3\u05E7\n');
testIt('ltr', 33, 'right', '\n Just\n <span>testing \u05E8\u05E7</span>\n ^\u05D1|\u05D5\u05D3\u05E7\n');
testIt('ltr', 34, 'right', '\n Just\n <span>testing \u05E8\u05E7</span>\n \u05D1^\u05D5|\u05D3\u05E7\n');
testIt('ltr', 35, 'right', '\n Just\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5^\u05D3|\u05E7\n');
testIt('ltr', 36, 'right', '\n Just\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3^\u05E7|\n');
testIt('ltr', 37, 'right', '\n Just\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7|\n');

// ltr left character
testIt('ltr', 0, 'left', '|\n Just\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 1, 'left', '\n| Just\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 2, 'left', '\n |Just\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 3, 'left', '\n |J^ust\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 4, 'left', '\n J|u^st\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 5, 'left', '\n Ju|s^t\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 6, 'left', '\n Jus|t^\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 7, 'left', '\n Just|\n^ <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 8, 'left', '\n Just|\n^ <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 9, 'left', '\n Just\n |<^span>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 10, 'left', '\n Just\n <s|pan>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</s|pan>');
testIt('ltr', 11, 'left', '\n Just\n <sp|an>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</sp|an>');
testIt('ltr', 12, 'left', '\n Just\n <spa|n>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</spa|n>');
testIt('ltr', 13, 'left', '\n Just\n <span|>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span|>');
testIt('ltr', 14, 'left', '\n Just|\n^ <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 15, 'left', '\n Just\n <span>|t^esting \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 16, 'left', '\n Just\n <span>t|e^sting \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 17, 'left', '\n Just\n <span>te|s^ting \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 18, 'left', '\n Just\n <span>tes|t^ing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 19, 'left', '\n Just\n <span>test|i^ng \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 20, 'left', '\n Just\n <span>testi|n^g \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 21, 'left', '\n Just\n <span>testin|g^ \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 22, 'left', '\n Just\n <span>testing| ^\u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 23, 'left', '\n Just\n <span>testing |\u05E8^\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 24, 'left', '\n Just\n <span>testing \u05E8|\u05E7^</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 25, 'left', '\n Just\n <span>testing \u05E8\u05E7|<^/span>\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('ltr', 26, 'left', '\n Just\n <span>testing \u05E8|\u05E7^span\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('ltr', 27, 'left', '\n Just\n <span>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('ltr', 28, 'left', '\n Just\n <span>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('ltr', 29, 'left', '\n Just\n <span>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('ltr', 30, 'left', '\n Just\n <span>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('ltr', 31, 'left', '\n Just\n <span>testing \u05E8|\u05E7^</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 32, 'left', '\n Just\n <span>testing \u05E8\u05E7</span>|\n^ \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 33, 'left', '\n Just\n <span>testing \u05E8\u05E7</span>|\n^ \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 34, 'left', '\n Just\n <span>testing \u05E8\u05E7</span>\n |\u05D1^\u05D5\u05D3\u05E7\n');
testIt('ltr', 35, 'left', '\n Just\n <span>testing \u05E8\u05E7</span>\n \u05D1|\u05D5^\u05D3\u05E7\n');
testIt('ltr', 36, 'left', '\n Just\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5|\u05D3^\u05E7\n');
testIt('ltr', 37, 'left', '\n Just\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3|\u05E7^\n');

// ltr forward character
testIt('ltr', 0, 'forward', '\n ^J|ust\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 1, 'forward', '\n ^J|ust\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 2, 'forward', '\n ^J|ust\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 3, 'forward', '\n J^u|st\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 4, 'forward', '\n Ju^s|t\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 5, 'forward', '\n Jus^t|\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 6, 'forward', '\n Just^\n| <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 7, 'forward', '\n Just\n <span>^t|esting \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 8, 'forward', '\n Just\n <span>^t|esting \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 9, 'forward', '\n Just\n <^s|pan>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 10, 'forward', '\n Just\n <s|pan>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</s|pan>');
testIt('ltr', 11, 'forward', '\n Just\n <sp|an>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</sp|an>');
testIt('ltr', 12, 'forward', '\n Just\n <spa|n>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</spa|n>');
testIt('ltr', 13, 'forward', '\n Just\n <span|>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span|>');
testIt('ltr', 14, 'forward', '\n Just\n <span>^t|esting \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 15, 'forward', '\n Just\n <span>t^e|sting \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 16, 'forward', '\n Just\n <span>te^s|ting \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 17, 'forward', '\n Just\n <span>tes^t|ing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 18, 'forward', '\n Just\n <span>test^i|ng \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 19, 'forward', '\n Just\n <span>testi^n|g \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 20, 'forward', '\n Just\n <span>testin^g| \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 21, 'forward', '\n Just\n <span>testing^ |\u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 22, 'forward', '\n Just\n <span>testing ^\u05E8|\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 23, 'forward', '\n Just\n <span>testing \u05E8^\u05E7|</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 24, 'forward', '\n Just\n <span>testing \u05E8\u05E7</span>^\n| \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 25, 'forward', '\n Just\n <span>testing \u05E8\u05E7<^/|span>\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('ltr', 26, 'forward', '\n Just\n <span>testing \u05E8\u05E7span^\n| \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('ltr', 27, 'forward', '\n Just\n <span>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('ltr', 28, 'forward', '\n Just\n <span>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('ltr', 29, 'forward', '\n Just\n <span>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('ltr', 30, 'forward', '\n Just\n <span>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('ltr', 31, 'forward', '\n Just\n <span>testing \u05E8\u05E7</span>^\n| \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 32, 'forward', '\n Just\n <span>testing \u05E8\u05E7</span>\n ^\u05D1|\u05D5\u05D3\u05E7\n');
testIt('ltr', 33, 'forward', '\n Just\n <span>testing \u05E8\u05E7</span>\n ^\u05D1|\u05D5\u05D3\u05E7\n');
testIt('ltr', 34, 'forward', '\n Just\n <span>testing \u05E8\u05E7</span>\n \u05D1^\u05D5|\u05D3\u05E7\n');
testIt('ltr', 35, 'forward', '\n Just\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5^\u05D3|\u05E7\n');
testIt('ltr', 36, 'forward', '\n Just\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3^\u05E7|\n');
testIt('ltr', 37, 'forward', '\n Just\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7|\n');

// ltr backward character
testIt('ltr', 0, 'backward', '|\n Just\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 1, 'backward', '\n| Just\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 2, 'backward', '\n |Just\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 3, 'backward', '\n |J^ust\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 4, 'backward', '\n J|u^st\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 5, 'backward', '\n Ju|s^t\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 6, 'backward', '\n Jus|t^\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 7, 'backward', '\n Just|\n^ <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 8, 'backward', '\n Just|\n^ <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 9, 'backward', '\n Just\n |<^span>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 10, 'backward', '\n Just\n <s|pan>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</s|pan>');
testIt('ltr', 11, 'backward', '\n Just\n <sp|an>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</sp|an>');
testIt('ltr', 12, 'backward', '\n Just\n <spa|n>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</spa|n>');
testIt('ltr', 13, 'backward', '\n Just\n <span|>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span|>');
testIt('ltr', 14, 'backward', '\n Just|\n^ <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 15, 'backward', '\n Just\n <span>|t^esting \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 16, 'backward', '\n Just\n <span>t|e^sting \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 17, 'backward', '\n Just\n <span>te|s^ting \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 18, 'backward', '\n Just\n <span>tes|t^ing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 19, 'backward', '\n Just\n <span>test|i^ng \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 20, 'backward', '\n Just\n <span>testi|n^g \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 21, 'backward', '\n Just\n <span>testin|g^ \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 22, 'backward', '\n Just\n <span>testing| ^\u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 23, 'backward', '\n Just\n <span>testing |\u05E8^\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 24, 'backward', '\n Just\n <span>testing \u05E8|\u05E7^</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 25, 'backward', '\n Just\n <span>testing \u05E8\u05E7|<^/span>\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('ltr', 26, 'backward', '\n Just\n <span>testing \u05E8|\u05E7^span\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('ltr', 27, 'backward', '\n Just\n <span>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('ltr', 28, 'backward', '\n Just\n <span>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('ltr', 29, 'backward', '\n Just\n <span>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('ltr', 30, 'backward', '\n Just\n <span>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('ltr', 31, 'backward', '\n Just\n <span>testing \u05E8|\u05E7^</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 32, 'backward', '\n Just\n <span>testing \u05E8\u05E7</span>|\n^ \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 33, 'backward', '\n Just\n <span>testing \u05E8\u05E7</span>|\n^ \u05D1\u05D5\u05D3\u05E7\n');
testIt('ltr', 34, 'backward', '\n Just\n <span>testing \u05E8\u05E7</span>\n |\u05D1^\u05D5\u05D3\u05E7\n');
testIt('ltr', 35, 'backward', '\n Just\n <span>testing \u05E8\u05E7</span>\n \u05D1|\u05D5^\u05D3\u05E7\n');
testIt('ltr', 36, 'backward', '\n Just\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5|\u05D3^\u05E7\n');
testIt('ltr', 37, 'backward', '\n Just\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3|\u05E7^\n');

// rtl right character
testIt('rtl', 0, 'right', '\n ^J|ust\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 1, 'right', '\n ^J|ust\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 2, 'right', '\n ^J|ust\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 3, 'right', '\n J^u|st\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 4, 'right', '\n Ju^s|t\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 5, 'right', '\n Jus^t|\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 6, 'right', '\n Just^\n| <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 7, 'right', '\n Just\n <span>^t|esting \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 8, 'right', '\n Just\n <span>^t|esting \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 9, 'right', '\n Just\n <^s|pan>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 10, 'right', '\n Just\n <s|pan>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</s|pan>');
testIt('rtl', 11, 'right', '\n Just\n <sp|an>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</sp|an>');
testIt('rtl', 12, 'right', '\n Just\n <spa|n>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</spa|n>');
testIt('rtl', 13, 'right', '\n Just\n <span|>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span|>');
testIt('rtl', 14, 'right', '\n Just\n <span>^t|esting \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 15, 'right', '\n Just\n <span>t^e|sting \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 16, 'right', '\n Just\n <span>te^s|ting \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 17, 'right', '\n Just\n <span>tes^t|ing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 18, 'right', '\n Just\n <span>test^i|ng \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 19, 'right', '\n Just\n <span>testi^n|g \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 20, 'right', '\n Just\n <span>testin^g| \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 21, 'right', '\n Just\n <span>testing^ |\u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 22, 'right', '\n Just\n <span>testing ^\u05E8|\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 23, 'right', '\n Just\n <span>testing \u05E8^\u05E7|</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 24, 'right', '\n Just\n <span>testing \u05E8\u05E7</span>^\n| \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 25, 'right', '\n Just\n <span>testing \u05E8\u05E7<^/|span>\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('rtl', 26, 'right', '\n Just\n <span>testing \u05E8\u05E7span^\n| \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('rtl', 27, 'right', '\n Just\n <span>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('rtl', 28, 'right', '\n Just\n <span>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('rtl', 29, 'right', '\n Just\n <span>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('rtl', 30, 'right', '\n Just\n <span>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('rtl', 31, 'right', '\n Just\n <span>testing \u05E8\u05E7</span>^\n| \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 32, 'right', '\n Just\n <span>testing \u05E8\u05E7</span>\n ^\u05D1|\u05D5\u05D3\u05E7\n');
testIt('rtl', 33, 'right', '\n Just\n <span>testing \u05E8\u05E7</span>\n ^\u05D1|\u05D5\u05D3\u05E7\n');
testIt('rtl', 34, 'right', '\n Just\n <span>testing \u05E8\u05E7</span>\n \u05D1^\u05D5|\u05D3\u05E7\n');
testIt('rtl', 35, 'right', '\n Just\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5^\u05D3|\u05E7\n');
testIt('rtl', 36, 'right', '\n Just\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3^\u05E7|\n');
testIt('rtl', 37, 'right', '\n Just\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7|\n');

// rtl left character
testIt('rtl', 0, 'left', '|\n Just\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 1, 'left', '\n| Just\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 2, 'left', '\n |Just\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 3, 'left', '\n |J^ust\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 4, 'left', '\n J|u^st\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 5, 'left', '\n Ju|s^t\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 6, 'left', '\n Jus|t^\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 7, 'left', '\n Just|\n^ <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 8, 'left', '\n Just|\n^ <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 9, 'left', '\n Just\n |<^span>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 10, 'left', '\n Just\n <s|pan>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</s|pan>');
testIt('rtl', 11, 'left', '\n Just\n <sp|an>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</sp|an>');
testIt('rtl', 12, 'left', '\n Just\n <spa|n>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</spa|n>');
testIt('rtl', 13, 'left', '\n Just\n <span|>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span|>');
testIt('rtl', 14, 'left', '\n Just|\n^ <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 15, 'left', '\n Just\n <span>|t^esting \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 16, 'left', '\n Just\n <span>t|e^sting \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 17, 'left', '\n Just\n <span>te|s^ting \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 18, 'left', '\n Just\n <span>tes|t^ing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 19, 'left', '\n Just\n <span>test|i^ng \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 20, 'left', '\n Just\n <span>testi|n^g \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 21, 'left', '\n Just\n <span>testin|g^ \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 22, 'left', '\n Just\n <span>testing| ^\u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 23, 'left', '\n Just\n <span>testing |\u05E8^\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 24, 'left', '\n Just\n <span>testing \u05E8|\u05E7^</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 25, 'left', '\n Just\n <span>testing \u05E8\u05E7|<^/span>\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('rtl', 26, 'left', '\n Just\n <span>testing \u05E8|\u05E7^span\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('rtl', 27, 'left', '\n Just\n <span>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('rtl', 28, 'left', '\n Just\n <span>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('rtl', 29, 'left', '\n Just\n <span>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('rtl', 30, 'left', '\n Just\n <span>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('rtl', 31, 'left', '\n Just\n <span>testing \u05E8|\u05E7^</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 32, 'left', '\n Just\n <span>testing \u05E8\u05E7</span>|\n^ \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 33, 'left', '\n Just\n <span>testing \u05E8\u05E7</span>|\n^ \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 34, 'left', '\n Just\n <span>testing \u05E8\u05E7</span>\n |\u05D1^\u05D5\u05D3\u05E7\n');
testIt('rtl', 35, 'left', '\n Just\n <span>testing \u05E8\u05E7</span>\n \u05D1|\u05D5^\u05D3\u05E7\n');
testIt('rtl', 36, 'left', '\n Just\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5|\u05D3^\u05E7\n');
testIt('rtl', 37, 'left', '\n Just\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3|\u05E7^\n');

// rtl forward character
testIt('rtl', 0, 'forward', '\n ^J|ust\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 1, 'forward', '\n ^J|ust\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 2, 'forward', '\n ^J|ust\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 3, 'forward', '\n J^u|st\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 4, 'forward', '\n Ju^s|t\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 5, 'forward', '\n Jus^t|\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 6, 'forward', '\n Just^\n| <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 7, 'forward', '\n Just\n <span>^t|esting \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 8, 'forward', '\n Just\n <span>^t|esting \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 9, 'forward', '\n Just\n <^s|pan>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 10, 'forward', '\n Just\n <s|pan>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</s|pan>');
testIt('rtl', 11, 'forward', '\n Just\n <sp|an>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</sp|an>');
testIt('rtl', 12, 'forward', '\n Just\n <spa|n>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</spa|n>');
testIt('rtl', 13, 'forward', '\n Just\n <span|>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span|>');
testIt('rtl', 14, 'forward', '\n Just\n <span>^t|esting \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 15, 'forward', '\n Just\n <span>t^e|sting \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 16, 'forward', '\n Just\n <span>te^s|ting \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 17, 'forward', '\n Just\n <span>tes^t|ing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 18, 'forward', '\n Just\n <span>test^i|ng \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 19, 'forward', '\n Just\n <span>testi^n|g \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 20, 'forward', '\n Just\n <span>testin^g| \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 21, 'forward', '\n Just\n <span>testing^ |\u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 22, 'forward', '\n Just\n <span>testing ^\u05E8|\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 23, 'forward', '\n Just\n <span>testing \u05E8^\u05E7|</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 24, 'forward', '\n Just\n <span>testing \u05E8\u05E7</span>^\n| \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 25, 'forward', '\n Just\n <span>testing \u05E8\u05E7<^/|span>\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('rtl', 26, 'forward', '\n Just\n <span>testing \u05E8\u05E7span^\n| \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('rtl', 27, 'forward', '\n Just\n <span>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('rtl', 28, 'forward', '\n Just\n <span>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('rtl', 29, 'forward', '\n Just\n <span>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('rtl', 30, 'forward', '\n Just\n <span>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('rtl', 31, 'forward', '\n Just\n <span>testing \u05E8\u05E7</span>^\n| \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 32, 'forward', '\n Just\n <span>testing \u05E8\u05E7</span>\n ^\u05D1|\u05D5\u05D3\u05E7\n');
testIt('rtl', 33, 'forward', '\n Just\n <span>testing \u05E8\u05E7</span>\n ^\u05D1|\u05D5\u05D3\u05E7\n');
testIt('rtl', 34, 'forward', '\n Just\n <span>testing \u05E8\u05E7</span>\n \u05D1^\u05D5|\u05D3\u05E7\n');
testIt('rtl', 35, 'forward', '\n Just\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5^\u05D3|\u05E7\n');
testIt('rtl', 36, 'forward', '\n Just\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3^\u05E7|\n');
testIt('rtl', 37, 'forward', '\n Just\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7|\n');

// rtl backward character
testIt('rtl', 0, 'backward', '|\n Just\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 1, 'backward', '\n| Just\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 2, 'backward', '\n |Just\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 3, 'backward', '\n |J^ust\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 4, 'backward', '\n J|u^st\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 5, 'backward', '\n Ju|s^t\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 6, 'backward', '\n Jus|t^\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 7, 'backward', '\n Just|\n^ <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 8, 'backward', '\n Just|\n^ <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 9, 'backward', '\n Just\n |<^span>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 10, 'backward', '\n Just\n <s|pan>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</s|pan>');
testIt('rtl', 11, 'backward', '\n Just\n <sp|an>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</sp|an>');
testIt('rtl', 12, 'backward', '\n Just\n <spa|n>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</spa|n>');
testIt('rtl', 13, 'backward', '\n Just\n <span|>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span|>');
testIt('rtl', 14, 'backward', '\n Just|\n^ <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 15, 'backward', '\n Just\n <span>|t^esting \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 16, 'backward', '\n Just\n <span>t|e^sting \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 17, 'backward', '\n Just\n <span>te|s^ting \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 18, 'backward', '\n Just\n <span>tes|t^ing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 19, 'backward', '\n Just\n <span>test|i^ng \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 20, 'backward', '\n Just\n <span>testi|n^g \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 21, 'backward', '\n Just\n <span>testin|g^ \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 22, 'backward', '\n Just\n <span>testing| ^\u05E8\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 23, 'backward', '\n Just\n <span>testing |\u05E8^\u05E7</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 24, 'backward', '\n Just\n <span>testing \u05E8|\u05E7^</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 25, 'backward', '\n Just\n <span>testing \u05E8\u05E7|<^/span>\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('rtl', 26, 'backward', '\n Just\n <span>testing \u05E8|\u05E7^span\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('rtl', 27, 'backward', '\n Just\n <span>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('rtl', 28, 'backward', '\n Just\n <span>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('rtl', 29, 'backward', '\n Just\n <span>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('rtl', 30, 'backward', '\n Just\n <span>testing \u05E8\u05E7\n \u05D1\u05D5\u05D3\u05E7\n</span>');
testIt('rtl', 31, 'backward', '\n Just\n <span>testing \u05E8|\u05E7^</span>\n \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 32, 'backward', '\n Just\n <span>testing \u05E8\u05E7</span>|\n^ \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 33, 'backward', '\n Just\n <span>testing \u05E8\u05E7</span>|\n^ \u05D1\u05D5\u05D3\u05E7\n');
testIt('rtl', 34, 'backward', '\n Just\n <span>testing \u05E8\u05E7</span>\n |\u05D1^\u05D5\u05D3\u05E7\n');
testIt('rtl', 35, 'backward', '\n Just\n <span>testing \u05E8\u05E7</span>\n \u05D1|\u05D5^\u05D3\u05E7\n');
testIt('rtl', 36, 'backward', '\n Just\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5|\u05D3^\u05E7\n');
testIt('rtl', 37, 'backward', '\n Just\n <span>testing \u05E8\u05E7</span>\n \u05D1\u05D5\u05D3|\u05E7^\n');
</script>
